git-svn-id: https://svn.qdecoder.org/qdecoder/trunk@132 5a34604f-89d6-db11-b6c0-0015c5f490bc

+++ src/qString.c
@@ -22,6 +22,9 @@
 #include <stdbool.h>
 #include <string.h>
 #include <stdarg.h>
+#include <unistd.h>
+#include <sys/types.h>
+#include <sys/time.h>
 #include "qDecoder.h"
 
 /**********************************************
@@ -659,7 +662,7 @@ char *qUniqId(void) {
 	nUsec = tv.tv_usec;
 #endif
 
-	sprintf(szSeed, "%ld-%ld.%ld-%s:%s", random(), time(NULL), nUsec, qGetenvDefault("", "REMOTE_ADDR"), qGetenvDefault("", "REMOTE_PORT"));
+	sprintf(szSeed, "%ld-%ld.%ld-%s:%s", random(), (long int)time(NULL), nUsec, qGetenvDefault("", "REMOTE_ADDR"), qGetenvDefault("", "REMOTE_PORT"));
 	strcpy(szUniqId, qMD5Str(szSeed));
 
 	return szUniqId;
